requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
asyncpg>=0.31.0
pydantic-settings>=2.12.0
python-dotenv>=1.2.1
